requirements.txt
pyvimeo==1.1.0
requests==2.25.0
urllib3==1.26.2
chardet==3.0.4
tuspy==0.2.4